+++ cmn-data-svc-client/src/main/java/org/acumos/cds/client/ICommonDataServiceRestClient.java
@@ -299,6 +299,42 @@ public interface ICommonDataServiceRestClient {
 			String[] validationStatusCodes, String[] tags, RestPageRequest pageRequest);
 
 	/**
+	 * Finds solutions editable by the specified user ('my models'). This includes
+	 * the user's private solutions and solutions co-owned by (shared with) the
+	 * user. This special-purpose method supports a dynamic search page on the
+	 * portal interface.
+	 * 
+	 * @param nameKeywords
+	 *            Keywords to perform "LIKE" search in Name field; ignored if null
+	 *            or empty
+	 * @param descriptionKeywords
+	 *            Keywords to perform "LIKE" search in Description field; ignored if
+	 *            null or empty
+	 * @param active
+	 *            Solution active status; true for active, false for inactive; required.
+	 * @param userId
+	 *            User ID who created a solution or has access to a solution; required.
+	 * @param accessTypeCodes
+	 *            Access type codes; use four-letter sequence "null" to match a null
+	 *            value; ignored if null or empty
+	 * @param modelTypeCodes
+	 *            Model type codes; use four-letter sequence "null" to match a null
+	 *            value; ignored if null or empty
+	 * @param validationStatusCodes
+	 *            Validation status codes; use four-letter sequence "null" to match
+	 *            a null value; ignored if null or empty
+	 * @param tags
+	 *            Solution tag names; ignored if null or empty
+	 * @param pageRequest
+	 *            Page index, page size and sort information; defaults to page 0 of
+	 *            size 20 if null.
+	 * @return Page of solution objects.
+	 */
+	RestPageResponse<MLPSolution> findUserSolutions(String[] nameKeywords, String[] descriptionKeywords, boolean active,
+			String userId, String[] accessTypeCodes, String[] modelTypeCodes, String[] validationStatusCodes,
+			String[] tags, RestPageRequest pageRequest);
+
+	/**
 	 * Searches the solutions.
 	 * 
 	 * @param queryParameters