+++ cmn-data-svc-server/src/test/java/org/acumos/cds/test/FOMRepositoryTest.java
@@ -23,14 +23,18 @@ import java.lang.invoke.MethodHandles;
 import java.util.Date;
 import java.util.List;
 
+import org.acumos.cds.AccessTypeCode;
+import org.acumos.cds.ValidationStatusCode;
 import org.acumos.cds.domain.MLPArtifact;
 import org.acumos.cds.domain.MLPSolRevArtMap;
+import org.acumos.cds.domain.MLPSolUserAccMap;
 import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPSolutionFOM;
 import org.acumos.cds.domain.MLPSolutionRevision;
 import org.acumos.cds.domain.MLPUser;
 import org.acumos.cds.repository.ArtifactRepository;
 import org.acumos.cds.repository.SolRevArtMapRepository;
+import org.acumos.cds.repository.SolUserAccMapRepository;
 import org.acumos.cds.repository.SolutionFOMRepository;
 import org.acumos.cds.repository.SolutionRepository;
 import org.acumos.cds.repository.SolutionRevisionRepository;
@@ -44,6 +48,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.test.context.junit4.SpringRunner;
 
 /**
@@ -66,6 +71,8 @@ public class FOMRepositoryTest {
 	@Autowired
 	private SolRevArtMapRepository solRevArtMapRepository;
 	@Autowired
+	private SolUserAccMapRepository solUserAccMapRepository;
+	@Autowired
 	private UserRepository userRepository;
 	@Autowired
 	private SolutionSearchService solutionSearchService;
@@ -77,10 +84,15 @@ public class FOMRepositoryTest {
 		boolean setupTeardown = true;
 
 		MLPUser cu = null;
+		MLPUser cu2 = null;
 		MLPSolution cs = null;
 		MLPSolutionRevision cr = null;
 		MLPArtifact ca = null;
 		MLPSolRevArtMap map = null;
+		MLPSolUserAccMap accMap = null;
+		final String name = "name";
+		final String accCode = AccessTypeCode.PR.name();
+		final String valCode = ValidationStatusCode.NV.name();
 
 		if (setupTeardown) {
 			// Create entities for query
@@ -91,11 +103,17 @@ public class FOMRepositoryTest {
 			Assert.assertNotNull(cu.getUserId());
 			logger.info("Created user {}", cu);
 
-			cs = new MLPSolution("sol name", cu.getUserId(), true);
+			final String loginName2 = "user2_" + Long.toString(new Date().getTime());
+			cu2 = new MLPUser(loginName2, "entityuser2@abc.com", true);
+			cu2 = userRepository.save(cu2);
+			Assert.assertNotNull(cu2.getUserId());
+			logger.info("Created user {}", cu2);
+
+			cs = new MLPSolution("some solution " + name, cu.getUserId(), true);
 			cs = solutionRepository.save(cs);
 			Assert.assertNotNull("Solution ID", cs.getSolutionId());
 
-			cr = new MLPSolutionRevision(cs.getSolutionId(), "version", cu.getUserId(), "PR", "NV");
+			cr = new MLPSolutionRevision(cs.getSolutionId(), "version", cu.getUserId(), accCode, valCode);
 			cr = revisionRepository.save(cr);
 			Assert.assertNotNull("Revision ID", cr.getRevisionId());
 			logger.info("Created solution revision {}", cr.getRevisionId());
@@ -106,7 +124,12 @@ public class FOMRepositoryTest {
 			logger.info("Created artifact {}", ca);
 
 			map = new MLPSolRevArtMap(cr.getRevisionId(), ca.getArtifactId());
-			solRevArtMapRepository.save(map);
+			map = solRevArtMapRepository.save(map);
+			logger.info("Created sol-rev-art map {}", map);
+
+			accMap = new MLPSolUserAccMap(cs.getSolutionId(), cu2.getUserId());
+			accMap = solUserAccMapRepository.save(accMap);
+			logger.info("Created sol-user-acc map {}", accMap);
 		}
 
 		// Find all via Spring repository
@@ -116,21 +139,38 @@ public class FOMRepositoryTest {
 		logger.info("Found FOM row count {}", foms.size());
 
 		// Find by modified date
-
-		String[] accTypes = new String[] { "PR" };
-		String[] valCodes = new String[] { "NV" };
+		String[] empty = new String[0];
+		String[] nameKw = new String[] { name }; // substring of solution name
+		String[] accTypes = new String[] { accCode };
+		String[] valCodes = new String[] { valCode };
 		Date modifiedDate = new Date();
 		modifiedDate.setTime(modifiedDate.getTime() - 60 * 1000);
 
 		// Via Hibernate constraint
-		logger.info("Querying for FOM via search service");
+		logger.info("Querying for FOM via search services");
+		Pageable pageable = new PageRequest(0, 6, null);
+
+		logger.info("Querying for FOM via findPortalSolutions method");
+		Page<MLPSolution> byName = solutionSearchService.findPortalSolutions(nameKw, empty, true, empty, empty,
+				accTypes, valCodes, empty, pageable);
+		Assert.assertTrue(byName != null && byName.getNumberOfElements() > 0);
+		logger.info("Found sols by name via criteria: size {}", byName.getContent().size());
+
 		Page<MLPSolution> solsByDate = solutionSearchService.findSolutionsByModifiedDate(true, accTypes, valCodes,
-				modifiedDate, new PageRequest(0, 6, null));
+				modifiedDate, pageable);
 		Assert.assertTrue(solsByDate != null && solsByDate.getNumberOfElements() > 0);
 		logger.info("Found sols by date via criteria: size {}", solsByDate.getContent().size());
 
+		// Find by user and Hibernate constraint - user2 owns no solutions but has
+		// access
+		Page<MLPSolution> byUser = solutionSearchService.findUserSolutions(nameKw, empty, true, cu2.getUserId(), empty,
+				empty, valCodes, empty, pageable);
+		Assert.assertTrue(byUser != null && byUser.getNumberOfElements() > 0);
+		logger.info("Found sols by user via criteria: size {}", byUser.getContent().size());
+
 		if (setupTeardown) {
 			// Clean up
+			solUserAccMapRepository.delete(accMap);
 			solRevArtMapRepository.delete(map);
 			artifactRepository.delete(ca);
 			revisionRepository.delete(cr);